Update patch set 1
Patch Set 1:

This will allow more use cases to leverage the TA platform without forcing the user to make changes to charts that they do not develop or maintain. The proposal is to include this backward compatibility option in Akraino Release 2 and removed it in Akraino Release 3 next spring.

Patch-set: 1