Update patch set 1
Patch Set 1:

I think 6 months is a good target, but the most important points are:
1) Communication of the intent to remove an API
2) Availability of a suitable replacement prior to the removal

I don't know if the replacements for these APIs were already present in the Akraino release 1 REC, but if they are then it would be reasonable to remove them as part of the release 2 work. But if so, it would have been good to communicate this as part of the scope of R2 and to have a Jira story for it. I'm pretty certain we did not have a Jira story for the removal of these APIs. I don't think we even had a Jira story for the Kubernetes upgrade.

Patch-set: 1