Bump minor RPM version to trigger new build
LF Nexus has expired old builds and there haven't been any new code
commits in a while because REC development has stopped as of Akraino
Release 4. REC will not be participating in Release 5, but this commit
is intended to trigger the CI/CD to create a new build so that anyone
who is interested can take a look at it.

Signed-off-by: Paul Carver <pcarver@att.com>
Change-Id: I29aa5d33594d056a591e08013552a7d6da2885b0